requirements.txt
numpy
PyOpenGL
PyOpenGL_accelerate
pygame